bookmarks_manager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.6
